responds to data changes now
